+++ src/env.h
@@ -25,8 +25,8 @@ MONGO_EXTERN_C_START
 
 /* This is a no-op in the generic implementation. */
 int mongo_env_set_socket_op_timeout( mongo *conn, int millis );
-int mongo_env_read_socket( mongo *conn, void *buf, int len );
-int mongo_env_write_socket( mongo *conn, const void *buf, int len );
+int mongo_env_read_socket( mongo *conn, void *buf, size_t len );
+int mongo_env_write_socket( mongo *conn, const void *buf, size_t len );
 int mongo_env_socket_connect( mongo *conn, const char *host, int port );
 
 /* Initialize socket services */
